Python - Update bindings
