+++ bindings/python/src/lib.rs
@@ -78,7 +78,6 @@ fn normalizers(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<normalizers::Sequence>()?;
     m.add_class::<normalizers::Lowercase>()?;
     m.add_class::<normalizers::Strip>()?;
-    m.add_class::<normalizers::ByteLevel>()?;
     Ok(())
 }
