+++ bindings/python/tokenizers/normalizers/__init__.py
@@ -9,7 +9,6 @@ NFKC = normalizers.NFKC
 Sequence = normalizers.Sequence
 Lowercase = normalizers.Lowercase
 Strip = normalizers.Strip
-ByteLevel = normalizers.ByteLevel
 
 
 NORMALIZERS = {"nfc": NFC, "nfd": NFD, "nfkc": NFKC, "nfkd": NFKD}
