+++ tokenizers/src/normalizers/mod.rs
@@ -2,6 +2,3 @@ pub mod bert;
 pub mod strip;
 pub mod unicode;
 pub mod utils;
-
-// Re-export these as normalizers
-pub use super::pre_tokenizers::byte_level;
